Added support for touchscreen and mobile devices
Added jQuery mobile so that ‘vmousemove’ could be used to allow mobile
touch screen devices to use the app as well.

Removed some of the inline styles and included bare minimum jQuery
mobile CSS to hide loadingmessage on page load.

Changed padding and margins so content would look better on mobile
devices, as well as changing how the footer works as it used to overlap
the option buttons at the bottom on mobile devices.